transferd user methods to account model
