+++ db/migrate/20240104143611_create_accounts.rb
@@ -1,0 +1,8 @@
+class CreateAccounts < ActiveRecord::Migration[7.1]
+  def change
+    create_table :accounts do |t|
+
+      t.timestamps
+    end
+  end
+end
