+++ test/models/account_test.rb
@@ -1,0 +1,7 @@
+require "test_helper"
+
+class AccountTest < ActiveSupport::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end
